refactor(parser): decouple lexer from diagnostics module
- Move lexer error/warning types into new lexer::error module
- Replace Arc<DiagnosticReporter> with take_errors/take_warnings drains
- Add From<LexError>/From<LexWarning> for Diagnostic as consumer bridge
- Drain lexer diagnostics from Parser::advance into the reporter
- Update lexer tests for accumulator API; add invalid-token coverage
