+++ crates/typhon-parser/src/diagnostics/mod.rs
@@ -15,17 +15,14 @@
 mod error;
 mod reporter;
 
-// Re-export public types
-pub use error::{
-    Diagnostic,
-    DiagnosticLevel,
-    LexError,
-    LexErrorBuilder,
-    LexErrorKind,
-    ParseErrorBuilder,
-    ParseResult,
-    ParseError,
-};
+// Re-export public types.
+//
+// Note: lexer error/warning types (`LexError`, `LexErrorBuilder`,
+// `LexErrorKind`, `LexWarning`) intentionally live in the `lexer` module so
+// that the lexer has no dependency on `diagnostics`. Consumers that need the
+// `Diagnostic` view of those types use the `From` impls defined in
+// [`error`] and access the source types via `crate::lexer`.
+pub use error::{Diagnostic, DiagnosticLevel, ParseError, ParseErrorBuilder, ParseResult};
 pub use reporter::{DiagnosticReporter, format_error_context, format_with_line_numbers};
 use typhon_source::types::SourceSpan;
